% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matrix.R
\name{calculate_matrix}
\alias{calculate_matrix}
\title{Calculate the matrix weight for a focal mean}
\usage{
calculate_matrix(Rast, Radius)
}
\arguments{
\item{Rast}{A raster (class: SpatRaster).}

\item{Radius}{A Radius in the units of the raster, to generate the weight.}
}
\value{
A square matrix with the weights to be used in a focal function
}
\description{
This function takes a raster and a desired radius and generates a circular
matrix with ones and NA to be later used in a focal function.
}
\examples{
# Load necessary libraries and create a sample categorical raster
library(terra)

data(dry_wet_nature)
DRY <- terra::unwrap(dry_wet_nature)
circle_matrix <- calculate_matrix(Rast = DRY, Radius = 2000)

plot(terra::rast(circle_matrix), colNA = "black")

}
\seealso{
\code{\link{SpatRaster}}, \code{\link{terra::focal}}
}
