% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prop.R
\name{calculate_prop}
\alias{calculate_prop}
\title{Calculate the proportion of each landuse in each cell of a raster}
\usage{
calculate_prop(Rast, Radius, verbose = TRUE)
}
\arguments{
\item{Rast}{A raster (class: SpatRaster).}

\item{Radius}{A Radius in the units of the raster, to generate the
proprotion.}

\item{verbose}{logical, if true (default), then messages of the progress is
written as messages in the function}
}
\value{
A stack of rasters with proportions, where each layer corresponds
to the proportion of each category of the raster.
}
\description{
This function takes a categorical raster and generates a stack of rasters
with and a desired radius and generates a stack of the proportion of each
landuse in each cell, considering only non NA cells.
}
\examples{
# Load necessary libraries and create a sample categorical raster
library(terra)

data(dry_wet_nature)
nature <- terra::unwrap(dry_wet_nature)
proportions <- calculate_prop(Rast = nature, Radius = 200)

plot(proportions, colNA = "black")

}
\seealso{
\link[terra:factors]{terra::focal}
}
