% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tempcont.R
\name{calculate_tempcont}
\alias{calculate_tempcont}
\title{Calculate the temporal continuity of each landuse in each cell of a raster}
\usage{
calculate_tempcont(Rast, years = NULL, Vars = NULL)
}
\arguments{
\item{Rast}{A raster stack (class: SpatRaster), ordered from the
oldest to the newest}

\item{years}{numeric vector of the same number of layers
corresponding to the years, default is NULL}

\item{Vars}{if not null, a character vector to select the
variables in which to use the function, see examples}
}
\description{
This function takes a stack of categorical rasters and generates a
layer of spatial continutiy of landuse in each cell, considering
only non NA cells.
The resulting output is a raster layer where each cell represents
the temporal continuity of land use. The function assesses
continuity by comparing consecutive layers of the raster stack.
Without specifying years, the continuity is based on changes in land
use categories, creating a categorical
scale where 0 indicates discontinuity and positive integers indicate
the duration of continuity. When specifying years, the function
considers the temporal aspect, calculating the difference in years
between consecutive layers and providing insights into the duration
of consistent land use over time.
}
\examples{
# Load necessary libraries and create a sample categorical raster
library(terra)

data(corine)
corine <- terra::unwrap(corine)
# see the dataset

plot(corine, legend = F, col = "black")

Tempcont <- calculate_tempcont(Rast = corine)

plot(Tempcont, colNA = "black")

# Use years

TempcontNum <- calculate_tempcont(Rast = corine,
                           years = c(1990, 2000, 2006, 2012, 2018
                           ))

plot(TempcontNum, colNA = "black")

# use years and vars

TempcontNum <- calculate_tempcont(Rast = corine,
                                 years = c(1990, 2000, 2006, 2012, 2018),
                                 Vars = c("Coniferous forest" ,"Mixed forest"))
}
