% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dry_wet_nature}
\alias{dry_wet_nature}
\title{Dry and wet nature in denmark}
\format{
\subsection{\code{dry_wet_nature}}{

A wrapped SpatRaster with a 10 by 10 meeter resolution:
\describe{
\item{Nature}{dry nature or wet nature}
\item{year}{Year}
...
}
}
}
\source{
\url{https://envs.au.dk/en/research-areas/society-environment-and-resources/land-use-and-gis/basemap}
}
\usage{
dry_wet_nature
}
\description{
A subset of data from basemap 4 raster in denmark, as it is stated in its
website "Basemap is a national map of land use and land cover for Denmark.
Basemap combines publicly available spatial information into one map. Basemap
is a raster map with a cell size of 10 x 10 meter." In our case we have a
subset of the 4th version of basemap which was generated in 2021 and shows
which cells are dry nature or wet nature in Denmark
}
\keyword{datasets}
