% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_stack.R
\name{generate_dummy_stack}
\alias{generate_dummy_stack}
\title{Generate Dummy Variable Stack from Categorical Raster}
\usage{
generate_dummy_stack(Rast, Vars = NULL)
}
\arguments{
\item{Rast}{A categorical raster (class: SpatRaster) representing landcover.}

\item{Vars}{if not null, a character vector to select the variables in
which to use the function, see examples}
}
\value{
A stack of rasters with dummy variables, where each layer
corresponds to a unique category in the input raster.
}
\description{
This function takes a categorical raster and generates a stack of rasters with
dummy variables, where each layer corresponds to one of the categorical variables.
}
\examples{
# Load necessary libraries and create a sample categorical raster
library(terra)

# Generate dummy variable stack
data("Landuse_DK")
Nature <- terra::unwrap(Landuse_DK)
dummy_stack <- generate_dummy_stack(Nature,
                                    Vars = c("Agriculture", "Forest"))

plot(dummy_stack, colNA = "black")

# Selecting only some variables

 dummy_stack2 <- generate_dummy_stack(Nature,
        Vars = c("Agriculture", "Forest"))
 plot(dummy_stack2, colNA = "black")
}
\seealso{
\link[terra:factors]{terra::levels}
}
