% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Landuse_DK}
\alias{Landuse_DK}
\title{Landuse in denmark}
\format{
\subsection{\code{dry_wet_nature}}{

A wrapped SpatRaster with a 10 by 10 meeter resolution:
\describe{
\item{Nature}{dry nature or wet nature}
\item{year}{Year}
...
}
}
}
\source{
\url{https://envs.au.dk/en/research-areas/society-environment-and-resources/land-use-and-gis/basemap}
}
\usage{
Landuse_DK
}
\description{
A subset of data from basemap 4 raster in denmark, as it is stated in its
website "Basemap is a national map of land use and land cover for Denmark.
Basemap combines publicly available spatial information into one map. Basemap
is a raster map with a cell size of 10 x 10 meter." In our case we have a
subset of the 4th version of basemap which was generated in 2021 and shows an
upscaled raster of 100 by 100 meter cells which show which cells are of class
Resource extraction, Unmapped, Agriculture, Built up, dry nature, Forest,
Recreation, Water body or wet nature in Denmark
}
\keyword{datasets}
