% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Polygons}
\alias{Polygons}
\title{Polygons}
\format{
\subsection{\code{Polygons}}{

A matrix with 3 columns:
\describe{
\item{id}{id of the polygon}
\item{x}{easting}
\item{y}{northing}
...
}
}
}
\usage{
Polygons
}
\description{
A matrix to be used as polygons
}
\examples{

library(terra)
data(Polygons)
data("Landuse_DK")
Landuse <- terra::unwrap(Landuse_DK)

v <- vect(Polygons, "polygons", crs = terra::crs(Landuse))
plot(v)
}
\keyword{datasets}
