% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_polygons.R
\name{summarise_polygons}
\alias{summarise_polygons}
\title{Calculate the proportion of each landuse in each polygon}
\usage{
summarise_polygons(
  Rast,
  Polygons,
  Vars,
  dist = NULL,
  type = "Inside",
  verbose = TRUE
)
}
\arguments{
\item{Rast}{A raster (class: SpatRaster).}

\item{Vars}{if not null, a character vector to select the variables in
which to use the function, see examples}

\item{dist}{number in meters, only works if using type "Outside" or "Both"}

\item{type}{a character value, on of "Inside", "Both" or "Outside", if value
is Inside, the proportions will be calculated only inside the polygons, if
Both is selected it will be calculated in the polygons plus a buffer of
distance dist in meters}

\item{verbose}{logical, if true (default), then messages of the progress is
written as messages in the function}
}
\value{
A dataframe with  the proportions, where each row corresponds to the
each shapefile, and each column to the proportion of each category of the
original raster.
}
\description{
This function takes a categorical raster and a polygon, and generates a
dataframe of the proportion of each Polygon in each cell, considering only
non NA cells.
}
\examples{
# Load necessary libraries and create a sample categorical raster
library(terra)
data(Polygons)
data("Landuse_DK")
Landuse <- terra::unwrap(Landuse_DK)

v <- vect(Polygons, "polygons", crs = terra::crs(Landuse))
Test <- summarise_polygons(Rast = Landuse, Polygons = v,
Vars = c("Agriculture","Forest"))

Test
# Example with type both

TestBoth <- summarise_polygons(Rast = Landuse, Polygons = v,
  Vars = c("Agriculture","Forest"), type = "Both", dist = 200)
TestBoth
}
\seealso{
\link[terra:factors]{terra::zonal}
}
