% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_stack.R
\name{generate_dummy_stack}
\alias{generate_dummy_stack}
\title{Generate Dummy Variable Stack from Categorical Raster}
\usage{
generate_dummy_stack(Rast)
}
\arguments{
\item{Rast}{A categorical raster (class: SpatRaster) representing landcover.}
}
\value{
A stack of rasters with dummy variables, where each layer
corresponds to a unique category in the input raster.
}
\description{
This function takes a categorical raster and generates a stack of rasters with
dummy variables, where each layer corresponds to one of the categorical variables.
}
\examples{
# Load necessary libraries and create a sample categorical raster
library(terra)

# Generate dummy variable stack
data(dry_wet_nature)
DRY <- terra::unwrap(dry_wet_nature)
dummy_stack <- generate_dummy_stack(DRY)

plot(dummy_stack, colNA = "black")

}
\seealso{
\link[terra:factors]{terra::levels}
}
